requirements.txt
Django==1.8.1
django-daterange-filter==1.3.0
django-extensions==1.7.5
django-grappelli==2.9.1
django-uuidfield==0.5.0
djangorestframework==3.1.0
djutil==0.2
gevent==1.2.1
gunicorn==19.6.0
grequests==0.3.0
gspread==0.6.2

jsonfield==1.0.3
pycrypto==2.6.1
logstash-formatter==0.5.16
oauth2client==1.5.2
psycopg2==2.6
psycogreen
pygments==2.2.0
requests==2.7.0

